*: Move some infoschema operations to `temptable` package (#27818)
