+++ executor/set.go
@@ -24,10 +24,10 @@ import (
 	"github.com/pingcap/parser/mysql"
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/expression"
-	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/plugin"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/sessionctx/variable"
+	"github.com/pingcap/tidb/table/temptable"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/collate"
 	"github.com/pingcap/tidb/util/gcutil"
@@ -270,13 +270,6 @@ func (e *SetExecutor) loadSnapshotInfoSchemaIfNeeded(snapshotTS uint64) error {
 		return err
 	}
 
-	if local := vars.LocalTemporaryTables; local != nil {
-		snapInfo = &infoschema.TemporaryTableAttachedInfoSchema{
-			InfoSchema:           snapInfo,
-			LocalTemporaryTables: local.(*infoschema.LocalTemporaryTables),
-		}
-	}
-
-	vars.SnapshotInfoschema = snapInfo
+	vars.SnapshotInfoschema = temptable.AttachLocalTemporaryTableInfoSchema(e.ctx, snapInfo)
 	return nil
 }
