+++ planner/core/logical_plan_builder.go
@@ -49,6 +49,7 @@ import (
 	"github.com/pingcap/tidb/statistics"
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/table/tables"
+	"github.com/pingcap/tidb/table/temptable"
 	"github.com/pingcap/tidb/types"
 	driver "github.com/pingcap/tidb/types/parser_driver"
 	util2 "github.com/pingcap/tidb/util"
@@ -3872,12 +3873,10 @@ func (b *PlanBuilder) buildDataSource(ctx context.Context, tn *ast.TableName, as
 
 	is := b.is
 	if len(b.buildingViewStack) > 0 {
-		if tempIs, ok := is.(*infoschema.TemporaryTableAttachedInfoSchema); ok {
-			// For tables in view, always ignore local temporary table, considering the below case:
-			// If a user created a normal table `t1` and a view `v1` referring `t1`, and then a local temporary table with a same name `t1` is created.
-			// At this time, executing 'select * from v1' should still return all records from normal table `t1` instead of temporary table `t1`.
-			is = tempIs.InfoSchema
-		}
+		// For tables in view, always ignore local temporary table, considering the below case:
+		// If a user created a normal table `t1` and a view `v1` referring `t1`, and then a local temporary table with a same name `t1` is created.
+		// At this time, executing 'select * from v1' should still return all records from normal table `t1` instead of temporary table `t1`.
+		is = temptable.DetachLocalTemporaryTableInfoSchema(is)
 	}
 
 	tbl, err := is.TableByName(dbName, tn.Name)
